Created class IntRand, finished Person and on Data maxTime is now final

+++ src/com/leonardomazzon/artistadistrada/Data.java
@@ -3,13 +3,13 @@ package com.leonardomazzon.artistadistrada;
 public class Data {
 	public Semaphore artist;
 	public Semaphore chairs;
-	public final int maxTime;	//in ms
+	public final long maxTime;	//in ms
 	
 	/**
 	 * Constructor of class Data
 	 * @param maxTime maximum time that a person can use to obtain a chair in ms
 	 */
-	public Data(int maxTime) {
+	public Data(long maxTime) {
 		this.maxTime = maxTime;
 		artist = new Semaphore(1);
 		chairs = new Semaphore(4);
